Modified LoadEventsController and LoadEventsInteractor to handle clicks on Events
